+++ mycodo/controller_input.py
@@ -324,7 +324,8 @@ class InputController(threading.Thread):
         elif self.device == 'DS18B20':
             from mycodo.inputs.ds18b20 import DS18B20Sensor
             self.measure_input = DS18B20Sensor(self.location,
-                                              convert_to_unit=self.convert_to_unit)
+                                               self.resolution,
+                                               convert_to_unit=self.convert_to_unit)
         elif self.device == 'DS18S20':
             from mycodo.inputs.ds18s20 import DS18S20Sensor
             self.measure_input = DS18S20Sensor(self.location,
@@ -332,14 +333,17 @@ class InputController(threading.Thread):
         elif self.device == 'DS1822':
             from mycodo.inputs.ds1822 import DS1822Sensor
             self.measure_input = DS1822Sensor(self.location,
+                                              self.resolution,
                                               convert_to_unit=self.convert_to_unit)
         elif self.device == 'DS28EA00':
             from mycodo.inputs.ds28ea00 import DS28EA00Sensor
             self.measure_input = DS28EA00Sensor(self.location,
+                                                self.resolution,
                                                 convert_to_unit=self.convert_to_unit)
         elif self.device == 'DS1825':
             from mycodo.inputs.ds1825 import DS1825Sensor
             self.measure_input = DS1825Sensor(self.location,
+                                              self.resolution,
                                               convert_to_unit=self.convert_to_unit)
         elif self.device == 'DHT11':
             from mycodo.inputs.dht11 import DHT11Sensor