+++ CHANGELOG.md
@@ -6,14 +6,16 @@
  - Add unit conversions: celsius, fahrenheit, kelvin, meters, feet
  - Add ability to select whether lowering PID outputs are stored as positive or negative values
  - Add Sunrise/Sunset Conditional ([#440](https://github.com/kizniche/mycodo/issues/440))
- - Add ability to set the precision of DS18B20 sensors ([#439](https://github.com/kizniche/mycodo/issues/439))
+ - Add ability to set the precision for DS18B20, DS1822, DS28EA00, and DS1825 sensors ([#439](https://github.com/kizniche/mycodo/issues/439))
  - Add Inputs: DS18S20, DS1822, DS28EA00, DS1825, MAX31850K
+ - Add Input option to select resolution for DS18B20, DS1822, DS28EA00, and DS1825 ([#439](https://github.com/kizniche/mycodo/issues/439))
 
 
 ### Bugfixes
 
  - Fix issues with PID control on Dashboard ([#441](https://github.com/kizniche/mycodo/issues/441))
  - Improve LCD controller shutdown speed
+ - Fix installer not displaying progress in console ([#442](https://github.com/kizniche/mycodo/issues/442))
 
 
 ## 5.6.10 (2018-03-31)